Given Wald statistics, compute p-values.

In the TESSERA algorithm, and in generalized linear mixed models in general, the exact distribution of p-values for Wald statistics under the null hypothesis is unknown. We know that the standard errors are, for finite samples, likely biased downward (too small). Moreover, for finite samples, due to the approximations in the E-step of the fitting algorithm, there may be a small positive bias. Hence, we use the fdrtool fndr method, which assumes that the null is a centered normal distribution with some variance (estimated). This function assumes that fit_empirical_null has been run; that function also provides p-values, but, if you want to fit a distribution on one set of statistics and then evaluate p-values on another set of statistics using the fitted distribution, this function can be used.

Usage

calc_Wald_pvalue_from_fdrtool(wald_stats, params)

Arguments

wald_stats

A numeric vector of Wald statistics. Note: This function expects signed t-statistics, or, the un-squared statistics.

params

The dataframe of parameters from fit_empirical_null or from fdrtool::fdrtool; has one row and a column called sd.

Value

Vector of p-values (un-adjusted for multiple corrections).

Author

Florica J Constantine, florica AT berkeley.edu

Examples

# Simulate null and non-null Wald statistics
null_stats <- rnorm(n = 2000, mean = 0.1, sd = 4)
alt_stats <- rnorm(n = 500, mean = 5, sd = 2)
wald_vec <- c(null_stats, alt_stats)

# Fit
emp_null <- fit_empirical_null_fdrtool(wald_vec)
#> Step 1... determine cutoff point
#> Step 2... estimate parameters of null distribution and eta0
#> Step 3... compute p-values and estimate empirical PDF/CDF
#> Step 4... compute q-values and local fdr
#> 
calc_Wald_pvalue_from_fdrtool(wald_vec, emp_null$params)
#>    [1] 0.28163023 0.82626889 0.05891706 0.98786244 0.61267886 0.35820761
#>    [7] 0.15945254 0.86180136 0.86372517 0.84008531 0.67889340 0.60860169
#>   [13] 0.10173549 0.20857412 0.67389574 0.15010234 0.69715578 0.98275714
#>   [19] 0.65650525 0.48633551 0.69939164 0.76130516 0.31639791 0.55033594
#>   [25] 0.13406006 0.95476442 0.47572386 0.99434628 0.53013244 0.24416488
#>   [31] 0.45207025 0.87464091 0.83336973 0.19676605 0.91454991 0.93198705
#>   [37] 0.13993563 0.84220897 0.82131473 0.39239544 0.94068204 0.63061534
#>   [43] 0.98440426 0.85924337 0.71298288 0.02947477 0.95533436 0.63698137
#>   [49] 0.91072656 0.13959271 0.48731094 0.86431796 0.88724432 0.97240570
#>   [55] 0.96592371 0.65264397 0.07822730 0.03400371 0.79234823 0.85194802
#>   [61] 0.38935257 0.61623631 0.37249792 0.86268071 0.36681175 0.45656212
#>   [67] 0.39337150 0.90236548 0.68755646 0.18982355 0.25762724 0.80872909
#>   [73] 0.28610636 0.83764307 0.29288953 0.66734224 0.62083932 0.94697581
#>   [79] 0.87729905 0.95920478 0.12692437 0.30009875 0.54459341 0.64903665
#>   [85] 0.68201689 0.37393418 0.04277577 0.91849178 0.71936134 0.77986567
#>   [91] 0.72505156 0.39422072 0.40050418 0.99181415 0.68898596 0.18370000
#>   [97] 0.79649098 0.44691445 0.29341697 0.83156472 0.77670573 0.55156162
#>  [103] 0.41917466 0.54628458 0.17522795 0.60228405 0.67611862 0.68868934
#>  [109] 0.84350160 0.43203794 0.88551561 0.28185182 0.82430548 0.74415534
#>  [115] 0.64208114 0.09200647 0.72465761 0.19386712 0.82987830 0.39000318
#>  [121] 0.10600589 0.71026580 0.26727022 0.72365490 0.91731713 0.97045952
#>  [127] 0.62257454 0.85238099 0.69353842 0.86498746 0.37999600 0.49514982
#>  [133] 0.93340477 0.31618107 0.47134357 0.70741779 0.51565878 0.83752137
#>  [139] 0.47131968 0.94238327 0.91415347 0.52986510 0.13651733 0.53545483
#>  [145] 0.43134920 0.29163817 0.39135491 0.67265372 0.24521276 0.22257195
#>  [151] 0.72214078 0.90830044 0.38348502 0.67642030 0.39882766 0.58217945
#>  [157] 0.96034521 0.79527050 0.52703078 0.51605886 0.14205537 0.45187199
#>  [163] 0.82399206 0.82154574 0.16682890 0.55008376 0.38562143 0.76088680
#>  [169] 0.83813763 0.67090081 0.95135699 0.45740591 0.31033640 0.44015237
#>  [175] 0.53427931 0.40613584 0.72522804 0.32757156 0.82608372 0.07987718
#>  [181] 0.42599991 0.37985825 0.19628351 0.22459847 0.65018551 0.11406379
#>  [187] 0.91946946 0.04260960 0.39470018 0.36055419 0.36836230 0.77083825
#>  [193] 0.53194378 0.66451081 0.16132505 0.90154611 0.75323584 0.36219654
#>  [199] 0.32144692 0.61788881 0.75152348 0.27800069 0.97135688 0.84654995
#>  [205] 0.05793561 0.94351603 0.40059357 0.63392772 0.11042756 0.03628679
#>  [211] 0.37694442 0.80479549 0.52523275 0.32944044 0.65698907 0.93752315
#>  [217] 0.64774720 0.73021473 0.26374589 0.44932363 0.80582699 0.96841216
#>  [223] 0.95928146 0.90693667 0.44593978 0.32446481 0.77477949 0.72592389
#>  [229] 0.56560820 0.23895392 0.83719708 0.31327997 0.54551036 0.22863485
#>  [235] 0.94004751 0.63038637 0.52073267 0.58348984 0.35617775 0.19327056
#>  [241] 0.49187323 0.29312663 0.37830271 0.33598432 0.27084641 0.27028797
#>  [247] 0.98242983 0.96696857 0.71495085 0.90406948 0.52837674 0.70785377
#>  [253] 0.36016509 0.56933770 0.24625059 0.52920198 0.80531872 0.72160944
#>  [259] 0.64982262 0.24489158 0.89120895 0.27014840 0.99081850 0.88344259
#>  [265] 0.49013585 0.10386068 0.13307551 0.46022818 0.95154758 0.83564722
#>  [271] 0.94637538 0.09189690 0.94203701 0.91439046 0.97945673 0.13626216
#>  [277] 0.08728625 0.56509903 0.53517481 0.82450412 0.41680514 0.48389867
#>  [283] 0.64501108 0.78556025 0.75908502 0.36586795 0.47299068 0.84919722
#>  [289] 0.25938395 0.77879526 0.90706989 0.87756896 0.44077226 0.69211621
#>  [295] 0.20201483 0.24333358 0.27607580 0.48012898 0.61126177 0.09422962
#>  [301] 0.79542562 0.41569967 0.38815863 0.34478482 0.86114495 0.76880088
#>  [307] 0.61564904 0.11750146 0.13484321 0.22083119 0.68681079 0.37018988
#>  [313] 0.64741429 0.16002698 0.74324063 0.95807750 0.34554962 0.85590043
#>  [319] 0.41935535 0.86090801 0.59684334 0.77268808 0.23939968 0.93944177
#>  [325] 0.09206782 0.25583405 0.73464486 0.26654033 0.76290029 0.59466334
#>  [331] 0.39899535 0.12591176 0.33417145 0.91005286 0.80213534 0.31195677
#>  [337] 0.38462671 0.38646582 0.91582011 0.49182860 0.86980014 0.09823722
#>  [343] 0.94636262 0.28415615 0.38165934 0.17564377 0.96560359 0.74683323
#>  [349] 0.94119278 0.52518155 0.66447634 0.92487505 0.66589133 0.16852046
#>  [355] 0.24854966 0.22692182 0.77863058 0.14699685 0.68015980 0.51033685
#>  [361] 0.80286090 0.39431874 0.51169364 0.15222697 0.95001010 0.43020560
#>  [367] 0.18618564 0.52745261 0.37817468 0.90723511 0.87320752 0.59798676
#>  [373] 0.46388766 0.45676461 0.80604483 0.35659779 0.73628962 0.72696992
#>  [379] 0.93245143 0.32705052 0.77926919 0.80088917 0.95407484 0.96267498
#>  [385] 0.74751781 0.97137324 0.97964373 0.45420109 0.60544467 0.77659294
#>  [391] 0.76179055 0.45258936 0.14413892 0.59451725 0.97906695 0.10688501
#>  [397] 0.30255476 0.35389524 0.88902479 0.78222341 0.16790809 0.52070619
#>  [403] 0.46335163 0.41434314 0.24938651 0.36380291 0.60791561 0.09593991
#>  [409] 0.64481066 0.20754205 0.99090458 0.62097306 0.61378911 0.72246826
#>  [415] 0.29247187 0.60445395 0.67280917 0.17645262 0.47197741 0.84600754
#>  [421] 0.63524554 0.90475268 0.54866960 0.44640225 0.02250968 0.56018473
#>  [427] 0.59821751 0.14332166 0.93677993 0.48084444 0.70771721 0.51793001
#>  [433] 0.64342662 0.35665299 0.56696272 0.71621941 0.37161710 0.72183492
#>  [439] 0.91249408 0.17231163 0.27898085 0.56538289 0.34109205 0.78806309
#>  [445] 0.70291715 0.81560264 0.32299641 0.43131090 0.22815973 0.77089688
#>  [451] 0.15768963 0.74721892 0.48407294 0.22209871 0.96311630 0.10673507
#>  [457] 0.98215373 0.79563923 0.35904874 0.87791228 0.41403519 0.85172482
#>  [463] 0.18742997 0.55751414 0.48151259 0.84447820 0.65627701 0.93927969
#>  [469] 0.34550154 0.60201165 0.63593130 0.53995271 0.51298822 0.09316234
#>  [475] 0.29018758 0.71395506 0.66293781 0.91132605 0.68907580 0.30560910
#>  [481] 0.85280159 0.59583734 0.54456029 0.47054950 0.23573958 0.61937215
#>  [487] 0.65974164 0.54562773 0.13247874 0.11105519 0.94410356 0.85084816
#>  [493] 0.74087307 0.27791759 0.70620865 0.84801679 0.40647920 0.76177777
#>  [499] 0.80780594 0.27684272 0.59088703 0.59053311 0.80449716 0.86141204
#>  [505] 0.35610007 0.99074331 0.57719609 0.56745620 0.42610309 0.35991532
#>  [511] 0.34186008 0.03587241 0.91773149 0.75502585 0.89200885 0.49583438
#>  [517] 0.46684543 0.93469384 0.33399841 0.41068402 0.58051859 0.89711270
#>  [523] 0.66452214 0.65242776 0.65272482 0.61929545 0.94854303 0.02932365
#>  [529] 0.48722589 0.55316053 0.61677823 0.61006683 0.70572612 0.81710578
#>  [535] 0.69978457 0.55756244 0.61692596 0.64249094 0.84742611 0.36666073
#>  [541] 0.15000414 0.96326184 0.27549529 0.63713477 0.88072804 0.60601124
#>  [547] 0.74243623 0.62098208 0.61123396 0.59961787 0.55250052 0.46739470
#>  [553] 0.15464630 0.76965577 0.82396021 0.54087633 0.91850985 0.80906320
#>  [559] 0.19184232 0.62795608 0.63178607 0.92178055 0.18683334 0.25425334
#>  [565] 0.04020502 0.43020375 0.59308291 0.59682897 0.65940017 0.81955745
#>  [571] 0.52356953 0.29771833 0.53487336 0.63324637 0.51023832 0.79743716
#>  [577] 0.15142082 0.49621874 0.98661342 0.72681907 0.69444706 0.95620546
#>  [583] 0.05307562 0.88056738 0.76874874 0.25338398 0.53572989 0.35043375
#>  [589] 0.93015316 0.70127850 0.22645511 0.81216408 0.98208908 0.67326819
#>  [595] 0.57349720 0.94753313 0.67480958 0.86754234 0.54456238 0.13858477
#>  [601] 0.83801041 0.60861738 0.72873217 0.11702482 0.70628580 0.39558377
#>  [607] 0.46951700 0.69435966 0.54646159 0.83909284 0.34844547 0.05721820
#>  [613] 0.25157928 0.74987951 0.47248090 0.93979991 0.28641546 0.51305568
#>  [619] 0.58817900 0.27417780 0.34941037 0.27999890 0.37765958 0.57132510
#>  [625] 0.51104324 0.40368933 0.99247100 0.51320823 0.71032315 0.44612716
#>  [631] 0.90697087 0.39721146 0.47082657 0.80573382 0.48006285 0.37884786
#>  [637] 0.38365583 0.26881541 0.71336456 0.49781851 0.52259568 0.91032050
#>  [643] 0.61754068 0.55905527 0.47437734 0.97612919 0.84459701 0.58098170
#>  [649] 0.34843099 0.73528397 0.56971562 0.91450118 0.76784081 0.83619834
#>  [655] 0.91913290 0.55956757 0.64632297 0.39691548 0.66114841 0.55122923
#>  [661] 0.34842515 0.41258833 0.70112579 0.52313636 0.93431055 0.47842258
#>  [667] 0.79015861 0.88861428 0.73154060 0.04975441 0.09153810 0.72699808
#>  [673] 0.79457764 0.62631505 0.98429376 0.72931043 0.63425720 0.35749731
#>  [679] 0.86909869 0.20516089 0.24703591 0.11273006 0.56855425 0.93189683
#>  [685] 0.41960487 0.08607226 0.06604092 0.43573213 0.95391034 0.55867099
#>  [691] 0.55166501 0.58275987 0.57992914 0.87257144 0.24473543 0.66175897
#>  [697] 0.11772886 0.73723224 0.65253783 0.96667501 0.47218605 0.78310934
#>  [703] 0.62131660 0.98742732 0.69786028 0.63067362 0.63241359 0.91793669
#>  [709] 0.34665182 0.71146885 0.07197006 0.89964549 0.12218181 0.75680529
#>  [715] 0.78199561 0.32933884 0.23520491 0.82322504 0.99694386 0.06403142
#>  [721] 0.47377951 0.25367295 0.71242639 0.27478261 0.99694501 0.94742345
#>  [727] 0.82057439 0.79680095 0.90130039 0.64710839 0.89933830 0.53760439
#>  [733] 0.20586893 0.67647335 0.16355094 0.62879276 0.41469146 0.90860009
#>  [739] 0.40962386 0.38111633 0.69274108 0.60734679 0.88949588 0.18071990
#>  [745] 0.42960174 0.91831597 0.98663065 0.46715252 0.72971181 0.91331296
#>  [751] 0.94449315 0.48376145 0.46825322 0.54514628 0.58727104 0.88778424
#>  [757] 0.90421511 0.23767381 0.85870390 0.58763910 0.76205486 0.26485005
#>  [763] 0.75549669 0.22769011 0.98332389 0.67294957 0.65260333 0.48450260
#>  [769] 0.57859943 0.91415121 0.16833010 0.42806110 0.50312304 0.62050829
#>  [775] 0.93934169 0.45312481 0.46349240 0.82641242 0.68365809 0.45442931
#>  [781] 0.68812023 0.73786931 0.74564171 0.64341456 0.71483026 0.69103509
#>  [787] 0.65698746 0.07604361 0.79287293 0.26315108 0.27226409 0.39539793
#>  [793] 0.60685429 0.38480450 0.27939115 0.76161970 0.08576375 0.89774157
#>  [799] 0.27253471 0.71493813 0.22192701 0.75936854 0.69800663 0.49297450
#>  [805] 0.77833674 0.52881854 0.87945142 0.23395545 0.83976435 0.96498193
#>  [811] 0.76434094 0.20084985 0.25463788 0.46969309 0.41514303 0.96082252
#>  [817] 0.17824735 0.45930215 0.87307469 0.45703812 0.03336088 0.26707450
#>  [823] 0.27813431 0.78981801 0.50873668 0.70612844 0.33169988 0.67702536
#>  [829] 0.10506150 0.96310360 0.49191577 0.54748997 0.64851329 0.33802726
#>  [835] 0.88613911 0.72517965 0.70604656 0.64605463 0.05640236 0.91273632
#>  [841] 0.74498087 0.66844403 0.13821298 0.12531344 0.66888642 0.52962252
#>  [847] 0.53768524 0.46374475 0.44578962 0.30022415 0.76654579 0.66372662
#>  [853] 0.82281996 0.86375230 0.83447427 0.38774132 0.38998447 0.10356370
#>  [859] 0.48857232 0.70506830 0.98649459 0.60752120 0.25171900 0.77352732
#>  [865] 0.17687723 0.99146136 0.54544549 0.48543100 0.93866825 0.32262986
#>  [871] 0.56052450 0.65850408 0.41715022 0.68634160 0.58889848 0.90407122
#>  [877] 0.88175291 0.58232739 0.68801269 0.65273579 0.34458007 0.86679608
#>  [883] 0.83015466 0.40141267 0.32219733 0.75405880 0.67704844 0.79766027
#>  [889] 0.84648080 0.46306738 0.95190345 0.46887532 0.77209639 0.44327471
#>  [895] 0.65921371 0.25570709 0.28949566 0.30109338 0.57738036 0.72744668
#>  [901] 0.27107186 0.18315113 0.34665488 0.92271617 0.90475147 0.51671165
#>  [907] 0.75960901 0.33256558 0.32179214 0.57789509 0.99887628 0.79333853
#>  [913] 0.82959593 0.29716185 0.62491832 0.97898109 0.92780522 0.62066831
#>  [919] 0.72059607 0.97329215 0.51346962 0.27895788 0.33156188 0.24686627
#>  [925] 0.96195986 0.64265948 0.87972052 0.30834916 0.06724342 0.27817130
#>  [931] 0.82682997 0.80162649 0.94893700 0.95151526 0.12163714 0.47925852
#>  [937] 0.31561437 0.92803689 0.46776219 0.65287332 0.50828379 0.99111963
#>  [943] 0.69921928 0.48742721 0.89081639 0.71557792 0.29235495 0.83161291
#>  [949] 0.44903251 0.80387853 0.84410228 0.74898362 0.24911099 0.16844247
#>  [955] 0.70620254 0.56627028 0.10193798 0.96945074 0.83198652 0.81607323
#>  [961] 0.44873611 0.99838943 0.07839569 0.88760510 0.50864673 0.36809115
#>  [967] 0.17438150 0.94795708 0.35164739 0.77645710 0.42017531 0.59129534
#>  [973] 0.53186780 0.71752082 0.75568171 0.20414311 0.13712742 0.75066030
#>  [979] 0.27274733 0.50527187 0.75941845 0.43034465 0.83428935 0.47471703
#>  [985] 0.23065702 0.41507775 0.88342127 0.82312738 0.25192228 0.25945951
#>  [991] 0.64550140 0.60145644 0.49795216 0.94286117 0.47276051 0.37171117
#>  [997] 0.69824344 0.48803461 0.97164457 0.64512337 0.80717349 0.88121544
#> [1003] 0.61290995 0.32422016 0.31808432 0.78298064 0.91970074 0.59519921
#> [1009] 0.23323929 0.82778791 0.28983249 0.47702180 0.58554736 0.32576133
#> [1015] 0.19087914 0.90158710 0.28896762 0.31789149 0.14719087 0.78778579
#> [1021] 0.09941981 0.77644622 0.22500632 0.89328889 0.46168768 0.38750458
#> [1027] 0.24722115 0.20232680 0.03211488 0.82075254 0.41118014 0.61949250
#> [1033] 0.77725141 0.86477223 0.84750686 0.25423460 0.55719625 0.06547104
#> [1039] 0.97475869 0.72768579 0.84142565 0.63858769 0.88749813 0.83875406
#> [1045] 0.91542566 0.81477033 0.21029916 0.64254780 0.97837876 0.92228752
#> [1051] 0.81965690 0.66478458 0.66422453 0.44885766 0.59368158 0.32754555
#> [1057] 0.55269911 0.75698024 0.85685066 0.78751428 0.33258170 0.74204544
#> [1063] 0.36221310 0.86598879 0.29357268 0.31884011 0.09817199 0.95553762
#> [1069] 0.57317726 0.72481451 0.16147943 0.10831338 0.72303743 0.67158009
#> [1075] 0.52652393 0.74033213 0.25637792 0.77177271 0.93403805 0.12209578
#> [1081] 0.44710958 0.72744521 0.97634954 0.96556725 0.99544974 0.58686148
#> [1087] 0.39271906 0.55426634 0.17012873 0.75244182 0.11624610 0.02974763
#> [1093] 0.22828657 0.96315996 0.54876038 0.32932588 0.91360199 0.64637953
#> [1099] 0.31740476 0.82931811 0.18820695 0.51555226 0.92520175 0.81462158
#> [1105] 0.90801404 0.34299358 0.14157897 0.95659902 0.96405438 0.70337612
#> [1111] 0.27080228 0.75920309 0.58043061 0.76045709 0.69928084 0.60584394
#> [1117] 0.50038050 0.95364674 0.86904611 0.64766651 0.58034472 0.03190942
#> [1123] 0.30339974 0.62678472 0.47736904 0.56014716 0.75616289 0.76579449
#> [1129] 0.45271313 0.85110462 0.96559645 0.27389106 0.37493893 0.38757292
#> [1135] 0.74507313 0.55536883 0.16020337 0.38946687 0.39534212 0.25501492
#> [1141] 0.23196886 0.52976432 0.39190494 0.90117689 0.21509180 0.88282064
#> [1147] 0.45434689 0.73292117 0.32582998 0.55316710 0.07796679 0.93671973
#> [1153] 0.24871360 0.26695184 0.24265016 0.57382200 0.82693568 0.77643899
#> [1159] 0.64831708 0.83222426 0.93165084 0.51082450 0.84607904 0.08718069
#> [1165] 0.38189511 0.54740827 0.35435066 0.78234999 0.69013030 0.30540550
#> [1171] 0.96631658 0.77163143 0.59992422 0.25066739 0.24996513 0.80093219
#> [1177] 0.99046711 0.31939940 0.30086314 0.66734708 0.59389949 0.15134985
#> [1183] 0.86201970 0.80787339 0.86053910 0.70440345 0.73335520 0.94760098
#> [1189] 0.57420011 0.09403008 0.87068870 0.48104020 0.45767878 0.70252984
#> [1195] 0.48173202 0.24991082 0.91213258 0.20822045 0.41563774 0.37987758
#> [1201] 0.48840980 0.01412467 0.41342065 0.59707369 0.89197331 0.54728562
#> [1207] 0.98428254 0.29474387 0.93073191 0.63578449 0.59438134 0.97227086
#> [1213] 0.57223895 0.11126612 0.98002376 0.83536845 0.99805714 0.52807263
#> [1219] 0.26368749 0.58422951 0.28849695 0.53809135 0.15412965 0.57738658
#> [1225] 0.92863813 0.78431731 0.93988256 0.83987539 0.11527592 0.46809211
#> [1231] 0.67606118 0.82596549 0.89220130 0.72235088 0.94022397 0.23655926
#> [1237] 0.70762580 0.48128796 0.31897889 0.40922465 0.38483795 0.97026001
#> [1243] 0.99778916 0.84249367 0.99950420 0.50568651 0.57015657 0.66035002
#> [1249] 0.22556262 0.45156928 0.11579458 0.15269940 0.48843974 0.13144963
#> [1255] 0.54382141 0.14869896 0.56994991 0.65767030 0.57446252 0.57370884
#> [1261] 0.97847136 0.14990646 0.32789121 0.16895842 0.70494371 0.18011296
#> [1267] 0.95910869 0.93810019 0.64536319 0.88441525 0.61774060 0.80830362
#> [1273] 0.61187729 0.40956737 0.38495432 0.46080619 0.33202469 0.60479833
#> [1279] 0.40232505 0.76015829 0.44147421 0.64972846 0.71808719 0.91869096
#> [1285] 0.85785712 0.45625884 0.74272330 0.41550284 0.60725336 0.71964925
#> [1291] 0.23977464 0.61527015 0.32604495 0.88639962 0.26239584 0.74119464
#> [1297] 0.78318906 0.26139658 0.54433123 0.69831434 0.37830662 0.84689147
#> [1303] 0.70562373 0.99497578 0.68584847 0.48059843 0.54873670 0.91282683
#> [1309] 0.56974110 0.33614226 0.42847620 0.99722702 0.84483955 0.68938819
#> [1315] 0.46077240 0.45207933 0.25095526 0.50840766 0.41617520 0.39836097
#> [1321] 0.97100025 0.44406738 0.77861241 0.75843802 0.72735456 0.19228726
#> [1327] 0.47106136 0.80335723 0.76398453 0.95430052 0.36572803 0.69390367
#> [1333] 0.69278058 0.25868443 0.96559289 0.77252410 0.66657464 0.54607958
#> [1339] 0.83250762 0.55476478 0.62739066 0.37462631 0.47228061 0.57007086
#> [1345] 0.75432098 0.38129861 0.83963554 0.89767585 0.19953330 0.12861949
#> [1351] 0.53522065 0.74898735 0.73420619 0.58862898 0.72190552 0.37002629
#> [1357] 0.29152808 0.27609626 0.86032573 0.75429330 0.37904497 0.58581512
#> [1363] 0.96505369 0.06091263 0.04092815 0.33264710 0.33699640 0.73914564
#> [1369] 0.45237900 0.72594307 0.20800123 0.62581976 0.70059224 0.57325270
#> [1375] 0.44099205 0.88937862 0.55999861 0.26684277 0.45996998 0.41497204
#> [1381] 0.66124369 0.45100423 0.34230485 0.69974667 0.85427895 0.81242344
#> [1387] 0.93856552 0.86147048 0.29089313 0.99035051 0.72660801 0.85917322
#> [1393] 0.19042167 0.60133002 0.44743311 0.43878586 0.20156396 0.39085742
#> [1399] 0.94754266 0.13190158 0.38390467 0.84191345 0.49616804 0.89924657
#> [1405] 0.57069766 0.67059833 0.89575499 0.96639244 0.76390780 0.99750412
#> [1411] 0.35876889 0.75496705 0.78504467 0.52857902 0.26637699 0.89624659
#> [1417] 0.81594074 0.40802846 0.07117589 0.68123081 0.97028143 0.40774928
#> [1423] 0.85085361 0.71324972 0.22874994 0.66418179 0.18320815 0.44475026
#> [1429] 0.32859087 0.38812779 0.64581731 0.67707790 0.70408615 0.23726139
#> [1435] 0.47760758 0.68866691 0.30481211 0.63132498 0.31491063 0.78902799
#> [1441] 0.33977715 0.46264262 0.55087095 0.76251146 0.69463154 0.86882754
#> [1447] 0.30882870 0.34847769 0.84082773 0.69076120 0.68946427 0.27191606
#> [1453] 0.31710640 0.60538584 0.42237427 0.80654653 0.96188396 0.18475170
#> [1459] 0.47898830 0.51138246 0.66591663 0.41682512 0.50250807 0.73037500
#> [1465] 0.61834891 0.65321782 0.68713069 0.90878763 0.56138802 0.64883923
#> [1471] 0.19038962 0.49768569 0.77612876 0.76612304 0.46889038 0.34116206
#> [1477] 0.98591951 0.52494542 0.35077182 0.84675218 0.82115742 0.83788542
#> [1483] 0.54217867 0.95656693 0.96936589 0.56257294 0.90667775 0.30075603
#> [1489] 0.39361651 0.52232851 0.24674445 0.83927673 0.95761899 0.98992086
#> [1495] 0.02204632 0.97166854 0.95413952 0.63062326 0.75368048 0.60549439
#> [1501] 0.66471538 0.91015904 0.13802357 0.23730038 0.39350003 0.20384686
#> [1507] 0.63023597 0.21265433 0.26479925 0.54838924 0.70733678 0.73822340
#> [1513] 0.43528359 0.66406822 0.20207783 0.84420859 0.70130458 0.53167295
#> [1519] 0.77838557 0.95723112 0.79667173 0.72368954 0.86550333 0.50611938
#> [1525] 0.23072237 0.92607281 0.78889352 0.93484353 0.73469971 0.66162200
#> [1531] 0.89506583 0.84814878 0.93842484 0.76471697 0.65230477 0.96040620
#> [1537] 0.21470603 0.42234938 0.10388040 0.87467312 0.36650248 0.71871018
#> [1543] 0.65320527 0.59853659 0.47930839 0.76856597 0.95324593 0.50033030
#> [1549] 0.84068542 0.54829767 0.98334938 0.36268247 0.76162362 0.68134028
#> [1555] 0.57400872 0.97760957 0.36020630 0.94076103 0.59273549 0.21266197
#> [1561] 0.68340153 0.29960300 0.21285700 0.33732312 0.20825046 0.31716040
#> [1567] 0.88519840 0.63066523 0.73995145 0.28446381 0.99212368 0.65612400
#> [1573] 0.16691864 0.81461612 0.74077441 0.72981674 0.31969128 0.57952448
#> [1579] 0.68278908 0.25144984 0.33457411 0.83653078 0.83022257 0.26970649
#> [1585] 0.59709569 0.13975831 0.15212799 0.29731331 0.82052000 0.43626636
#> [1591] 0.64564632 0.83619195 0.23031461 0.75846025 0.99689810 0.45463656
#> [1597] 0.28132631 0.24972098 0.75503091 0.57892635 0.59253148 0.24016351
#> [1603] 0.49573442 0.30170793 0.80024007 0.76564527 0.92899308 0.87704532
#> [1609] 0.16525873 0.96458558 0.90703563 0.74389668 0.75209405 0.88100818
#> [1615] 0.35617302 0.97517139 0.97099263 0.53035654 0.95479310 0.83372300
#> [1621] 0.33446680 0.96440039 0.87222960 0.93624352 0.64781423 0.91155884
#> [1627] 0.87372039 0.83399271 0.47929889 0.22776183 0.53795398 0.29542860
#> [1633] 0.51043080 0.90958569 0.47769176 0.16487906 0.46551269 0.72512730
#> [1639] 0.15066876 0.82726128 0.88210422 0.06631535 0.98234730 0.58644941
#> [1645] 0.84368927 0.65714365 0.76580543 0.25977695 0.83462853 0.26920858
#> [1651] 0.82064076 0.72132256 0.50007903 0.98291493 0.44429441 0.23413115
#> [1657] 0.80262883 0.59943045 0.23722762 0.62475604 0.70740772 0.76491203
#> [1663] 0.73284843 0.07945502 0.23646190 0.79604522 0.48040831 0.88192198
#> [1669] 0.10918480 0.15557740 0.87683270 0.56323123 0.39188209 0.33826648
#> [1675] 0.43078227 0.34073230 0.60649277 0.66227100 0.53355824 0.91623495
#> [1681] 0.99533449 0.77463573 0.76738330 0.30808829 0.50766327 0.72763856
#> [1687] 0.17388928 0.08629503 0.90037690 0.58568308 0.54649353 0.22750500
#> [1693] 0.33208023 0.23128153 0.37020512 0.61717944 0.23740453 0.74298618
#> [1699] 0.81818354 0.34612945 0.69338543 0.18448169 0.99284328 0.74466670
#> [1705] 0.55209744 0.66654513 0.93765292 0.94737093 0.49412064 0.04173825
#> [1711] 0.92560087 0.69835100 0.53134400 0.32642987 0.13567357 0.93743856
#> [1717] 0.56542722 0.70453074 0.91827682 0.90180013 0.45739907 0.33709381
#> [1723] 0.98667100 0.36274124 0.31266576 0.68173841 0.57965398 0.47867142
#> [1729] 0.45689457 0.99867736 0.71188854 0.81178720 0.91392195 0.39940085
#> [1735] 0.78285557 0.52488824 0.71487406 0.43624822 0.91662473 0.20667856
#> [1741] 0.02491413 0.43970988 0.67618160 0.82370459 0.55787328 0.16326931
#> [1747] 0.87347203 0.69497652 0.73919276 0.93909176 0.97105245 0.09680522
#> [1753] 0.29153012 0.33724779 0.55342498 0.74434668 0.86259964 0.75961089
#> [1759] 0.99429431 0.80141057 0.87799172 0.86772434 0.74637016 0.96019918
#> [1765] 0.60571800 0.48561827 0.75080282 0.97624384 0.95239158 0.65897350
#> [1771] 0.66975756 0.06043949 0.54070100 0.71889858 0.62286997 0.83332159
#> [1777] 0.84358623 0.89939052 0.39243885 0.98324046 0.89714569 0.69797491
#> [1783] 0.06120456 0.19061196 0.57954993 0.27761909 0.13054765 0.61197234
#> [1789] 0.41639658 0.19402791 0.65773148 0.24592647 0.13502267 0.50299669
#> [1795] 0.54514531 0.50014039 0.83267175 0.49988125 0.92260030 0.55937769
#> [1801] 0.42200574 0.66501422 0.35201755 0.87670933 0.63686502 0.94319070
#> [1807] 0.46672461 0.99474386 0.48231296 0.41765362 0.16764542 0.41476381
#> [1813] 0.38699186 0.74684120 0.31048913 0.20431928 0.93845858 0.45855806
#> [1819] 0.39184119 0.06677134 0.19424567 0.89975877 0.68191320 0.28747289
#> [1825] 0.99872212 0.89625919 0.74485712 0.13347114 0.25928659 0.62657136
#> [1831] 0.62295578 0.91945458 0.74306200 0.69934563 0.85640076 0.19248135
#> [1837] 0.22796473 0.67437443 0.23562794 0.76139804 0.53324226 0.32231779
#> [1843] 0.58998425 0.95168455 0.62589880 0.78509606 0.93239155 0.46066092
#> [1849] 0.85465181 0.58253280 0.75704359 0.40116580 0.23349653 0.42402872
#> [1855] 0.07841898 0.81936281 0.25583610 0.95292904 0.57226417 0.85791367
#> [1861] 0.68248048 0.43856703 0.35929469 0.88209368 0.18317988 0.91186571
#> [1867] 0.25571329 0.86651614 0.39754424 0.78531042 0.97080161 0.78848645
#> [1873] 0.45947937 0.25313457 0.45045126 0.74146165 0.66953807 0.56152632
#> [1879] 0.97743978 0.07179423 0.49443100 0.61461201 0.93599259 0.56309610
#> [1885] 0.66603258 0.72718551 0.61968968 0.53195220 0.53375060 0.93545753
#> [1891] 0.79005502 0.96890210 0.40126135 0.43744475 0.44992413 0.21743831
#> [1897] 0.85490783 0.36943816 0.62346901 0.40427815 0.59053133 0.29491959
#> [1903] 0.70954487 0.25183450 0.70317417 0.98561143 0.15684101 0.40815183
#> [1909] 0.66347611 0.21951582 0.71634561 0.56881877 0.58880450 0.55144063
#> [1915] 0.55398867 0.69326999 0.35254077 0.64135277 0.52802689 0.92437573
#> [1921] 0.22898863 0.57111863 0.96053477 0.83888860 0.41262712 0.24015376
#> [1927] 0.98538275 0.65555710 0.72515814 0.81345824 0.36345435 0.45588172
#> [1933] 0.60657820 0.99529713 0.23816709 0.96673900 0.60364393 0.49586971
#> [1939] 0.69070424 0.63354837 0.33242089 0.47217057 0.49465547 0.64994470
#> [1945] 0.37247657 0.97273095 0.78174681 0.83259934 0.60170494 0.45617053
#> [1951] 0.15938142 0.18650134 0.96694230 0.19896217 0.30041672 0.86407779
#> [1957] 0.72662737 0.70986967 0.21912182 0.25883408 0.39969232 0.66135921
#> [1963] 0.69443420 0.93930783 0.66028319 0.86146811 0.51765619 0.04585289
#> [1969] 0.81730547 0.50002788 0.55991683 0.70758350 0.59352125 0.46918533
#> [1975] 0.41483154 0.49998102 0.58707097 0.10974602 0.17827082 0.89734512
#> [1981] 0.72623852 0.61732360 0.61751677 0.91377094 0.44850520 0.35109967
#> [1987] 0.98609213 0.89593602 0.60686761 0.92084683 0.97488910 0.47433420
#> [1993] 0.97792742 0.77207756 0.48677513 0.48128819 0.78580443 0.83085762
#> [1999] 0.21042111 0.12569757 0.36089073 0.30497615 0.47171923 0.44069800
#> [2005] 0.38937640 0.45751039 0.39300762 0.27158745 0.19655535 0.09609083
#> [2011] 0.27897174 0.27006126 0.35729887 0.26920111 0.39431049 0.50154031
#> [2017] 0.33121291 0.36149863 0.36058149 0.20834135 0.28388957 0.94465642
#> [2023] 0.31263865 0.41743670 0.40493699 0.33646715 0.12666572 0.38925852
#> [2029] 0.54806639 0.10147634 0.49984492 0.28211791 0.54842964 0.56702879
#> [2035] 0.57121134 0.36902655 0.13597627 0.23804876 0.66255700 0.55735960
#> [2041] 0.27276188 0.25009931 0.46896085 0.97491398 0.27719963 0.37150207
#> [2047] 0.25282987 0.23792877 0.55077323 0.19349670 0.11850080 0.23559717
#> [2053] 0.08601442 0.20695774 0.35050386 0.19891296 0.09206419 0.40803007
#> [2059] 0.04116204 0.14209270 0.83157256 0.21505266 0.44216359 0.15620929
#> [2065] 0.12736622 0.30645678 0.27935557 0.16978342 0.33671959 0.46945498
#> [2071] 0.55106841 0.38049265 0.71089083 0.33653293 0.21352132 0.23457789
#> [2077] 0.22236557 0.47164744 0.13288699 0.58842007 0.21246008 0.45946940
#> [2083] 0.39978766 0.15482140 0.11428480 0.15794466 0.26551278 0.85664472
#> [2089] 0.43212225 0.33960082 0.25001230 0.61487810 0.24040462 0.57902521
#> [2095] 0.25260962 0.37266450 0.43369950 0.58620075 0.18197065 0.37499440
#> [2101] 0.19067897 0.14057581 0.48351991 0.40077953 0.57919154 0.17469178
#> [2107] 0.42916221 0.56787279 0.29055251 0.09907367 0.23202758 0.52305515
#> [2113] 0.30897675 0.79455691 0.12571000 0.25954829 0.33074256 0.71712551
#> [2119] 0.40049716 0.22163396 0.25403028 0.34725876 0.36705086 0.22172990
#> [2125] 0.33697488 0.58246033 0.22520298 0.50823520 0.19379265 0.15817172
#> [2131] 0.12486171 0.23990833 0.63636218 0.19399815 0.48120737 0.40233252
#> [2137] 0.47669730 0.59350767 0.44853462 0.65228007 0.46910625 0.97826247
#> [2143] 0.26761608 0.23882727 0.43295971 0.54250827 0.60926594 0.13157462
#> [2149] 0.46026039 0.47291772 0.63641297 0.19330904 0.03105019 0.15650043
#> [2155] 0.89974383 0.10697804 0.23803345 0.23975563 0.48963671 0.58591321
#> [2161] 0.20785553 0.25526725 0.69518947 0.61727121 0.12434687 0.23467670
#> [2167] 0.42884082 0.20903114 0.32982247 0.66461837 0.22116997 0.21303698
#> [2173] 0.06287414 0.79674580 0.34520145 0.46990120 0.40457830 0.20573152
#> [2179] 0.29802235 0.16993910 0.38850194 0.37530425 0.38118067 0.39349709
#> [2185] 0.08280961 0.90271142 0.21013748 0.40572959 0.26700814 0.38767636
#> [2191] 0.32554985 0.14767833 0.42624347 0.18105280 0.48247286 0.19814950
#> [2197] 0.25624538 0.46618491 0.39843356 0.58854434 0.22812978 0.30627582
#> [2203] 0.24969255 0.84551640 0.16557704 0.56065414 0.43472984 0.31059630
#> [2209] 0.63626028 0.76322466 0.15499684 0.17612107 0.48441505 0.35356043
#> [2215] 0.17669646 0.24480042 0.26706600 0.70977921 0.46199045 0.72916448
#> [2221] 0.18922063 0.08575446 0.52272539 0.20822840 0.22272876 0.27211648
#> [2227] 0.53774661 0.13078459 0.28982613 0.13801541 0.37058616 0.24214929
#> [2233] 0.57497164 0.60544153 0.06814054 0.41496907 0.31102581 0.44374076
#> [2239] 0.35142612 0.09223909 0.50425778 0.40221440 0.22020888 0.74323095
#> [2245] 0.43087230 0.33173241 0.17882969 0.37472166 0.56220619 0.74828387
#> [2251] 0.06121693 0.56820429 0.20852760 0.14520473 0.26686685 0.38459127
#> [2257] 0.31313089 0.40092865 0.12364730 0.32718450 0.26552105 0.51331695
#> [2263] 0.24300689 0.47823576 0.58183771 0.19086251 0.12160978 0.30184513
#> [2269] 0.26863724 0.14488280 0.57113528 0.08927235 0.91627708 0.45598131
#> [2275] 0.37425879 0.09564260 0.24402123 0.19873582 0.59611370 0.17677476
#> [2281] 0.39889442 0.99173834 0.12912412 0.05134777 0.64894734 0.07652487
#> [2287] 0.18841339 0.20624698 0.09987265 0.65245402 0.31134267 0.28973538
#> [2293] 0.55153958 0.22568018 0.16585396 0.63258705 0.39339179 0.80603926
#> [2299] 0.31034050 0.20851958 0.34945711 0.57067283 0.52884542 0.46151098
#> [2305] 0.28073676 0.41980373 0.12029464 0.19200095 0.53459614 0.38319027
#> [2311] 0.20308666 0.34279079 0.30004105 0.39366515 0.75506165 0.49599374
#> [2317] 0.49549513 0.27425698 0.26750525 0.37456662 0.50966157 0.28762393
#> [2323] 0.59411203 0.30132030 0.47181558 0.44157315 0.67037834 0.14558834
#> [2329] 0.25206363 0.35164098 0.21728071 0.24173094 0.20953720 0.19560718
#> [2335] 0.36372599 0.16187827 0.34392078 0.16190917 0.32810370 0.52080810
#> [2341] 0.20350557 0.13108598 0.25622609 0.21950309 0.80398108 0.45245968
#> [2347] 0.15942227 0.44117816 0.21256771 0.47716834 0.10624255 0.46451681
#> [2353] 0.65361392 0.25834397 0.54856547 0.40186803 0.45374926 0.10735124
#> [2359] 0.23930483 0.32305842 0.27378454 0.13939253 0.28968721 0.29286058
#> [2365] 0.20889009 0.26020804 0.25533981 0.26771657 0.10899756 0.23167000
#> [2371] 0.61338906 0.37322513 0.39827184 0.76502517 0.29704592 0.27328131
#> [2377] 0.29481912 0.66831675 0.37106420 0.43424129 0.10142216 0.76158531
#> [2383] 0.37892478 0.72600408 0.15262982 0.62911615 0.37841023 0.44264181
#> [2389] 0.45961270 0.19531171 0.13094094 0.17049337 0.62698895 0.66542047
#> [2395] 0.26842888 0.44901445 0.33551829 0.16970136 0.65919942 0.26773812
#> [2401] 0.15372820 0.66152310 0.51679095 0.79507294 0.16527420 0.28777742
#> [2407] 0.19632485 0.20815588 0.39278277 0.92266698 0.11686657 0.41526831
#> [2413] 0.36958448 0.16757196 0.87844610 0.22771012 0.25174860 0.27132875
#> [2419] 0.39965048 0.24450040 0.20988565 0.70814012 0.59426209 0.26104494
#> [2425] 0.28176939 0.14309020 0.40468963 0.18201757 0.20018557 0.26173580
#> [2431] 0.28444336 0.51188354 0.19723955 0.19429560 0.46659395 0.29866158
#> [2437] 0.64383790 0.44741520 0.63254509 0.35998775 0.61427459 0.12961059
#> [2443] 0.31196678 0.62897459 0.15029619 0.21084805 0.19636271 0.55761157
#> [2449] 0.47543575 0.20730399 0.17552511 0.35667129 0.11686590 0.21340250
#> [2455] 0.38168784 0.42406138 0.09704445 0.33024335 0.50428428 0.28071331
#> [2461] 0.52561647 0.43417313 0.18081637 0.41446364 0.43303977 0.30446600
#> [2467] 0.55548221 0.24240345 0.44545092 0.29649493 0.30954907 0.17838839
#> [2473] 0.50035298 0.39906666 0.30025538 0.47598381 0.46021627 0.06559106
#> [2479] 0.39530299 0.09599681 0.54960556 0.75102056 0.69004848 0.05374467
#> [2485] 0.37383527 0.15718362 0.41929585 0.14049247 0.12327162 0.53585314
#> [2491] 0.81872572 0.30594599 0.11640138 0.32945575 0.35467123 0.39693267
#> [2497] 0.26321145 0.26611206 0.02842816 0.21352174