Check inputs for the TESSERA_spNNGP method. Make sure that the input object has everything needed to run without error.

Check inputs for the TESSERA_spNNGP method. Make sure that the input object has everything needed to run without error.

Usage

check_inputs_TESSERA_spNNGP(TESSERAData_obj)

Arguments

TESSERAData_obj

Object containing data. Created by the prep_data method.

Value

Nothing.

Note

Does not return anything.

This method can be used to check a hand-created input object. E.g., if a user does not want to use prep_data.

Author

Florica J Constantine, florica AT berkeley.edu