Accurate tiling of spatial single-cell data

An algorithm for segmenting single-cell resolution spatial omics data into small multicellular tiles whose edges track with natural tissue boundaries. These tiles can then be used in downstream analysis to label and define tissue regions across samples.

See also

Author

Maintainer: Ilya Korsunsky ilya.korsunsky@gmail.com (ORCID)

Authors:

Daniel Stein daniel_stein@g.harvard.edu (ORCID)